requirements.txt
docx==0.2.4
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==10.0.0
pycparser==2.21
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
PyPDF2==3.0.1
pypdfium2==4.18.0
pytesseract==0.3.10
python-docx==0.8.11
bnlp_toolkit==4.0.0

